+++ src/SafeRegex/Exception/PregMalformedPatternException.php
@@ -1,6 +1,8 @@
 <?php
 namespace TRegx\SafeRegex\Exception;
 
+use TRegx\Exception\MalformedPatternException;
+
 class PregMalformedPatternException extends CompilePregException implements MalformedPatternException
 {
 }