+++ test/Functional/SafeRegex/ExceptionFactory/ExceptionFactoryTest.php
@@ -2,8 +2,8 @@
 namespace Test\Functional\TRegx\SafeRegex\ExceptionFactory;
 
 use PHPUnit\Framework\TestCase;
+use TRegx\Exception\MalformedPatternException;
 use TRegx\SafeRegex\Errors\ErrorsCleaner;
-use TRegx\SafeRegex\Exception\MalformedPatternException;
 use TRegx\SafeRegex\Exception\RuntimePregException;
 use TRegx\SafeRegex\Exception\SuspectedReturnPregException;
 use TRegx\SafeRegex\ExceptionFactory;