+++ test/Functional/SafeRegex/Guard/GuardedExecutionTest.php
@@ -5,9 +5,9 @@ use Exception;
 use PHPUnit\Framework\TestCase;
 use Test\Utils\Functions;
 use Test\Warnings;
+use TRegx\Exception\MalformedPatternException;
 use TRegx\SafeRegex\Errors\Errors\EmptyHostError;
 use TRegx\SafeRegex\Errors\ErrorsCleaner;
-use TRegx\SafeRegex\Exception\MalformedPatternException;
 use TRegx\SafeRegex\Exception\RuntimePregException;
 use TRegx\SafeRegex\Guard\GuardedExecution;